Add resource check before destroying

+++ src/commands/destroy.js
@@ -5,6 +5,7 @@ const deleteDLQ = require('../aws/deleteDLQ');
 const deleteEventSourceMapping = require('../aws/deleteEventSourceMapping');
 const terminateEC2Instance = require('../aws/terminateEC2Instance');
 const deleteSecurityGroup = require('../aws/deleteSecurityGroup');
+const { doesAPIResourceExist } = require('../aws/doesResourceExist');
 
 module.exports = async function destroy(resourceName, options, homedir) {
   const preLambda = `${resourceName}PreLambda`;
@@ -12,24 +13,30 @@ module.exports = async function destroy(resourceName, options, homedir) {
   const securityGroupEC2 = `${resourceName}EC2SecurityGroup`;
   const securityGroupPostLambda = `${resourceName}PostLambdaSecurityGroup`;
 
-  console.log('Waiting for EC2 instance to terminate')
-  await terminateEC2Instance(resourceName);
-  console.log('EC2 instance terminated. EBS volume persists for data preservation');
+  const resourceExists = await doesAPIResourceExist(resourceName, homedir);
 
-  await deleteSecurityGroup(securityGroupEC2);
-  await deleteSecurityGroup(securityGroupPostLambda);
-  console.log('Security groups deleted');
-  await deleteEventSourceMapping(resourceName);
+  if (resourceExists) {
+    console.log('Waiting for EC2 instance to terminate')
+    await terminateEC2Instance(resourceName);
+    console.log('EC2 instance terminated. EBS volume persists for data preservation');
 
-  await deleteDLQ(resourceName, homedir);
-  console.log('DLQ deleted');
-  await deleteSQS(resourceName, homedir);
-  console.log('SQS deleted');
+    await deleteSecurityGroup(securityGroupEC2);
+    await deleteSecurityGroup(securityGroupPostLambda);
+    console.log('Security groups deleted');
+    await deleteEventSourceMapping(resourceName);
 
-  await deleteLambda(preLambda);
-  await deleteLambda(postLambda);
-  console.log('Lambda functions deleted');
+    await deleteDLQ(resourceName, homedir);
+    console.log('DLQ deleted');
+    await deleteSQS(resourceName, homedir);
+    console.log('SQS deleted');
 
-  await deleteApiResource(resourceName, homedir);
-  console.log(`API Gateway endpoint ${resourceName} deleted`);
+    await deleteLambda(preLambda);
+    await deleteLambda(postLambda);
+    console.log('Lambda functions deleted');
+
+    await deleteApiResource(resourceName, homedir);
+    console.log(`API Gateway endpoint ${resourceName} deleted`);
+  } else {
+    console.log(`${resourceName} endpoint does not exist on AWS`);
+  }
 };
